@@ -1,5 +1,5 @@
 module github.com/MetaFFI/lang-plugin-go/go-runtime
 
-go 1.16
+go 1.18
 
-require github.com/MetaFFI/plugin-sdk v0.0.0-20240225125309-dab77e763d09
+require github.com/MetaFFI/plugin-sdk v0.0.0-20240228070827-8e1cbbbc984e
